klient/machine: rename client/testutil package to client/clienttest
