+++ go/src/koding/klient/machine/client/clienttest/server.go
@@ -1,4 +1,4 @@
-package testutil
+package clienttest
 
 import (
 	"sync"
